requirements.txt
Django==4.2
djangorestframework==3.15.0
django-cors-headers==4.0.0
scikit-learn==1.3.2
pandas==2.2.2
joblib==1.3.2
openrouteservice==2.3.3